+++ src/transport/grpc/mod.rs
@@ -74,9 +74,10 @@ pub enum GrpcError {
     /// The gRPC connection returned an error.
     #[error("Transport error while working with grpc: {0}")]
     Transport(#[from] tonic::transport::Error),
-    /// Status code error.
-    #[error("Status error {0}")]
-    Status(#[from] tonic::Status),
+    /// Status code error that is not [`Ok`](tonic::Code::Ok)
+    // NOTE: the `Status` struct is to big to return as an erro, so we box it
+    #[error("recived an error status code with {0}")]
+    Status(Box<Status>),
     #[error("Error while serializing the interfaces")]
     /// Couldn't serialize interface to json.
     InterfacesSerialization(#[from] serde_json::Error),
@@ -92,6 +93,12 @@ pub enum GrpcError {
     /// Failed to convert a proto message.
     #[error(transparent)]
     MessageHubProtoConversion(#[from] MessageHubProtoError),
+}
+
+impl From<Status> for GrpcError {
+    fn from(value: Status) -> Self {
+        Self::Status(Box::new(value))
+    }
 }
 
 type MessageHubClientWithInterceptor =